add statistics service. migrate from injectable to service annotation
