+++ src/app/services/golf-course-api.service.ts
@@ -1,12 +1,10 @@
 import { HttpClient, HttpHeaders } from '@angular/common/http';
-import { inject, Injectable } from '@angular/core';
+import { inject, Service } from '@angular/core';
 import { Observable } from 'rxjs';
 import { environment } from '../../environments/environment';
 import { CourseSearch } from '../models/generated/model/courseSearch';
 
-@Injectable({
-  providedIn: 'root',
-})
+@Service()
 export class GolfCourseApiService {
   private readonly http = inject(HttpClient);
