+++ src/app/services/round.service.ts
@@ -1,12 +1,10 @@
-import { Injectable, inject } from '@angular/core';
+import { inject, Service } from '@angular/core';
 import { Round, RoundVariety } from '../models/round';
 import { RoundVarietyScoresPipe } from '../pipes/round-variety-scores.pipe';
 import { AppStateService } from './app-state.service';
 import { LocalStorageService } from './local-storage.service';
 
-@Injectable({
-  providedIn: 'root',
-})
+@Service()
 export class RoundService {
   private readonly localStorageService = inject(LocalStorageService);
   private readonly appStateService = inject(AppStateService);
