+++ src/environments/environment.dev.ts
@@ -1,5 +1,5 @@
 // DO NOT COMMIT THIS FILE WITH YOUR API KEY.
 export const environment = {
   production: false,
-  golfCourseApiKey: '', // Set via build environment variable
+  golfCourseApiKey: 'QYWC6U2UPGFPBWFFH3AGH3UZYE', // Set via build environment variable
 };
